# Improved kayak_font