--- a/bevy_kayak_ui/src/render/font/extract.rs
+++ b/bevy_kayak_ui/src/render/font/extract.rs
@@ -21,29 +21,22 @@ pub fn extract_texts(
     _dpi: f32,
 ) -> Vec<ExtractQuadBundle> {
     let mut extracted_texts = Vec::new();
-    let (background_color, text_layout, layout, font, properties) =
-        match render_primitive {
-            RenderPrimitive::Text {
-                color,
-                text_layout,
-                layout,
-                font,
-                properties,
-                ..
-            } => (
-                color,
-                text_layout,
-                layout,
-                font,
-                *properties,
-            ),
-            _ => panic!(""),
-        };
+    let (background_color, text_layout, layout, font, properties) = match render_primitive {
+        RenderPrimitive::Text {
+            color,
+            text_layout,
+            layout,
+            font,
+            properties,
+            ..
+        } => (color, text_layout, layout, font, *properties),
+        _ => panic!(""),
+    };
 
     let font_handle = font_mapping.get_handle(font.clone()).unwrap();
     let font = match fonts.get(font_handle.clone()) {
         Some(font) => font,
-        None => return Vec::new()
+        None => return Vec::new(),
     };
 
     let base_position = Vec2::new(layout.posx, layout.posy + properties.font_size);