--- a/kayak_core/src/render_primitive.rs
+++ b/kayak_core/src/render_primitive.rs
@@ -1,4 +1,3 @@
-use morphorm::layout;
 use kayak_font::{TextLayout, TextProperties};
 use crate::{
     color::Color,