## Objective

Prepare for text selection vis-a-vis #89. Additionally, improve the code quality and performance of font logic.

The main change was updating the text measurement logic to be more consistent and return `TextLayout` so that we can get as much layout information as possible in a single method call.

## Changelog

### Added

* `TextProperties` - A struct used to contain details for laying out text (font size, alignment, etc.).
* `Line` - A struct that represents a single line of text, marking the beginning index and length for each character, glyph, and grapheme.
* `TextLayout` - A struct containing layout information for text. 
* Utility text-related functions and constants.
* Added `unicode-segmentation` dependency for splitting text into valid [graphemes](https://unicode.org/reports/tr29/).
* Added `xi_unicode` dependency for splitting text into [breakable segments](https://unicode.org/reports/tr14/).
* Added `KayakFont::missing_glyph` - Contains a glyph which can be rendered in place of a missing one.
* Added `KayakFont::max_glyph_size` - Contains the size of the largest character in the atlas.

### Removed

* Removed `KayakFont::get_layout(...)` - Consolidated text measurements into `KayakFont::measure(...)` which is now used by both extraction and layout logic.

### Changed

* Lots of changes to `KayakFont`:
  * `KayakFont::get_word_width(...)` now calculates tabs and spaces properly. Additionally, it now uses the glyph's advance instead of its width.
  * General improvements to `KayakFont::measure(...)`.
  * Improved word wrapping.
  * Trailing spaces are now trimmed from lines when right-aligned.
  * Improved readability and documentation.
* Updated `kayak_font` example renderer to support Bevy 0.7.
* Renamed `LayoutRect` to `GlyphRect`.
* `RenderPrimitive` now contains a cached `TextLayout` (in my tests, this saves us ~5 FPS in debug).
* Increased number of glyphs present in the example Roboto font (i.e. beyond just ASCII).
* Updated `kayak_font/examples/bevy.rs` example with controllable text box for easy testing.
* Renamed `Atlas::size` to `Atlas::font_size` for clarity.
* Increased `MAX_CHARACTERS` from `100` to `500`.

## Screenshots

#### Normal Layout
> Notice that, before, we didn't render tabs properly.

| Before | After |
| :----: | :---: |
| <img width="398" alt="normal-before" src="https://user-images.githubusercontent.com/49806985/164120860-bca8c927-4445-4d2b-a99c-778d8e946731.png"> | <img width="398" alt="normal-after" src="https://user-images.githubusercontent.com/49806985/164120859-7fe84a6a-1e9b-4e91-bfdc-c01647e6595f.png"> |

#### Right-Aligned Layout
> Notice that, before, we have spaces on the end that otherwise break up the clean alignment.

| Before | After |
| :----: | :---: |
| <img width="400" alt="right-before" src="https://user-images.githubusercontent.com/49806985/164120862-594e4dcc-5e83-4e2e-a51b-bfb8450cf3dc.png"> | <img width="399" alt="right-after" src="https://user-images.githubusercontent.com/49806985/164120861-217d7cbe-dd24-4a71-9fe1-9f55cb07859c.png"> |

#### Condensed Layout
> Notice that, before, chained words do not break at appropriate points.

| Before | After |
| :----: | :---: |
| <img width="229" alt="condensed-before" src="https://user-images.githubusercontent.com/49806985/164120858-02cbcdac-cbd1-48d8-ba3e-6588a9038da3.png"> | <img width="211" alt="condensed-after" src="https://user-images.githubusercontent.com/49806985/164120856-8dee4382-b44e-42eb-ae24-28e49464d63a.png"> |